Move argument parsing to args module
